#2 #4 - Redirect on first start

+++ implementation/gps-sleeper/src/pt/isec/gps/g22/sleeper/core/Profile.java
@@ -6,7 +6,7 @@ package pt.isec.gps.g22.sleeper.core;
 public class Profile {
 	private int id;
 	private int gender;
-	private int dateOfBirth;
+	private long dateOfBirth;
 	private int firstHourOfTheDay;
 
 	public static final String TABLE_NAME = "profile";
@@ -40,11 +40,11 @@ public class Profile {
 		this.gender = gender;
 	}
 
-	public int getDateOfBirth() {
+	public long getDateOfBirth() {
 		return dateOfBirth;
 	}
 
-	public void setDateOfBirth(int dateOfBirth) {
+	public void setDateOfBirth(long dateOfBirth) {
 		this.dateOfBirth = dateOfBirth;
 	}
